requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.1
django-filter==23.5
Pillow==10.2.0
reportlab==4.1.0
django-rest-knox==4.2.0
django-debug-toolbar==4.3.0 